+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Allow video and source tags.
+- Allow style attribute for divs.
+
 ## [2.19.0] - 2022-05-20
 
 ### Fixed